+++ programs/amm/src/prelude.rs
@@ -2,3 +2,4 @@ pub use crate::err::{self, AmmError};
 pub use crate::models::*;
 pub use crate::{consts, endpoints};
 pub use anchor_lang::prelude::*;
+pub use decimal::{Decimal, TryAdd, TryDiv, TryMul};